properly display shipping estimate with totals full mode
Magento won't calculate a shipping amount until the shipping method is
selected and a shipping address is entered. This is problematic when
totals full mode is enabled and we want to show an initial estimate.

Added an observer that sets a default shipping method when totals full
mode is enabled. This shipping method is configurable from the backend.
Estimation requires a country ID to be set, we set this value using
getDefaultCountry for now, which will go through our GeoService.

Closes #2.