update order summary when switching between shipping methods
Normally, Magento doesn't update the user's shipping information until
we navigate away from shipping page. There was no reason to do this,
because it never showed a full order summary in the first place.

Since we always force totals mode, we have to save shipping information
when switching between shipping methods. This ensures that we get an
accurate shipping cost estimation before we get to the review page.

Closes #15